auth/requirements.txt
amqp==2.6.0
anyjson==0.3.3
apturl==0.5.2
arrow==0.14.5
asgiref==3.2.10
atomicwrites==1.3.0
attrs==19.3.0
awscli==1.11.13
Babel==2.8.0
beautifulsoup4==4.4.1
billiard==3.6.3.0
blinker==1.3
boto==2.49.0
boto3==1.13.20
botocore==1.16.20
Brlapi==0.6.4
celery==4.3.0
certifi==2020.4.5.1
chardet==3.0.4
checkbox-support==0.22
codacy-coverage==1.3.11
colorama==0.3.7
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==1.2.3
cssselect==1.1.0
defer==1.0.6
defusedxml==0.6.0
Django==3.1.1
django-appconf==1.0.4
django-cleanup==4.0.0
django-cors-headers==3.5.0
django-crispy-forms==1.7.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.8
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==2.8
jmespath==0.10.0
kombu==4.6.9
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
oauthlib==3.1.0
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==6.2.1
plainbox==0.25
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
PyJWT==1.7.1
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.8
python-dateutil==2.8.1
python-debian==0.1.27
python-dotenv==0.15.0
python-http-client==3.3.1
python-memcached==1.53
python-systemd==231
python3-openid==3.1.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.11
reportlab==3.3.0
requests==2.22.0
requests-oauthlib==1.3.0
roman==2.0.0
rsa==3.2.3
s3transfer==0.3.3
sendgrid==6.4.8
sessioninstaller==0.0.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
uritemplate==3.0.1
urllib3==1.25.9
usb-creator==0.3.0
vine==1.3.0
virtualenv==15.0.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zipp==3.1.0
